feat: centralize slot navigation handling
